Refactored ASG start and stop functionality
