+++ package.json
@@ -6,9 +6,9 @@
   "scripts": {
     "deploy": "node_modules/serverless/bin/serverless deploy --verbose",
     "destroy": "node_modules/serverless/bin/serverless remove --verbose",
-    "lint": "eslint **/*.js",
+    "lint": "eslint *.js",
     "test": "npm run lint && npm run test.unit",
-    "test.unit": "mocha test/*.test.js"
+    "test.unit": "mocha test/**/*.test.js"
   },
   "repository": {
     "type": "git",
