+++ src/asgs/hasTag.js
@@ -1,0 +1,3 @@
+module.exports = function hasTag(asg, target) {
+  return asg.Tags.some(tag => tag.Key === target);
+};
