+++ src/utils/valueForKey.js
@@ -1,0 +1,3 @@
+module.exports = function valueForKey(tags, key) {
+  return tags.find(tag => tag.Key === key).Value;
+};
