+++ test/asgs/responses/emptyResponse.js
@@ -1,0 +1,2 @@
+const emptyResponse = { AutoScalingGroups: [] };
+module.exports = emptyResponse;
