+++ test/asgs/responses/paginatedStart.js
@@ -1,0 +1,6 @@
+const startOnePageResponse = require('./startOnePageResponse');
+const startTwoPageResponse = require('./startTwoPageResponse');
+
+module.exports = function paginatedStart(nextToken) {
+  return nextToken ? startOnePageResponse : startTwoPageResponse;
+};
