+++ test/asgs/responses/paginatedStop.js
@@ -1,0 +1,6 @@
+const stopOnePageResponse = require('./stopOnePageResponse');
+const stopTwoPageResponse = require('./stopTwoPageResponse');
+
+module.exports = function paginatedStop(nextToken) {
+  return nextToken ? stopOnePageResponse : stopTwoPageResponse;
+};
